Deepen regex hierarchy
